Moving renderDots into one separately file
